separate population_script into different modules
