+++ population/populate_db.py
@@ -1,0 +1,10 @@
+from population.setup import *
+from population.add import *
+from population.create_dummy import *
+from population.save import *
+from population.dummy_populate import dummy_populate
+from population.testadd import run_test
+
+def populate():
+    # run_test(resetafter=True)
+    dummy_populate()
